requirements.txt
fastapi ~= 0.103
uvicorn[standard]
httpx
openai
python-dotenv
